Another Attempt at Heroic Stats Implementation (#56)
* Another Attempt at Heroic Stats Implementation

Added Rules to tune these effects, as well as lesser version of several effects for pets based upon their owner's stats, because pets cannot have Heroic stats. Values below are default.

Heroic STR - Straight 3% increase to melee damage per point of Heroic Strength. 3% felt like a lot, but looking at itemization it lines up pretty well with growth of Spell Damage. If I understood the code paths correctly, this does affect pets at 1/3 rate, but this needs testing.

Heroic STA - 10 points of damage reduction, up to a cap of 50% of melee damage, per point of Heroic Stamina. Pets get 2/3 of this effect.

Heroic DEX - 0.25% increase to critical damage per point of Heroic Dexterity. Pets get 1/3 of this effect.

Heroic AGI - Allows a chance to quasi-flurry. I do not believe that this works for pets. The math here is a little wonky and works differently for melee and bow attacks. Does not work for throwing attacks at all.

In the case of melee attacks, each point of Heroic Agility gives a 1% chance to get an additional attack. If that attack hits, there are additional chances for more attacks, as long as they continue to hit, with decreasing probability. This simulates the intense agility of the character giving them extra opportunities to land hits.

In the case of ranged attacks, the player will attempt X number of additional shots, where X grows as Heroic Agility grows. This simulates the player firing multiple arrows at once.

Heroic INT - 0.25% increase to critical effectiveness (Damage, DoT, Healing) per point of Heroic Intelligence. Pets get the FULL effect here, due to technical restrictions (too disruptive for future maintainability to forward the data about the caster to this point in the code path). It may be possible to counter this by adding a negative modifier earlier in the code path. This is probably irrelevant as I don't think pets have any way to get any spell critical chance.

Heroic WIS- 10 points of damage reduction, up to a cap of 50% of spell damage, per point of Heroic Wisdom. Pets get 2/3 of this effect.

Co-Authored-By: mute <natanx@gmail.com>